+++ lib/client.js
@@ -17,6 +17,7 @@ var Room = require("./models/room");
 var User = require("./models/user");
 var webRtcCall = require("./webrtc/call");
 var utils = require("./utils");
+var contentRepo = require("./content-repo");
 
 var CRYPTO_ENABLED = false;
 
@@ -1404,67 +1405,6 @@ MatrixClient.prototype.setAvatarUrl = function(url, callback) {
 };
 
 /**
- * Get the avatar URL for a room member. <strong>This method is experimental and
- * may change.</strong>
- * @param {module:room-member.RoomMember} member
- * @param {Number} width The desired width of the thumbnail.
- * @param {Number} height The desired height of the thumbnail.
- * @param {string} resizeMethod The thumbnail resize method to use, either
- * "crop" or "scale".
- * @param {Boolean} allowDefault (optional) Passing false causes this method to
- * return null if the user has no avatar image. Otherwise, a default image URL
- * will be returned.
- * @return {?string} the avatar URL or null.
- */
-MatrixClient.prototype.getAvatarUrlForMember =
-        function(member, width, height, resizeMethod, allowDefault) {
-    if (!member || !member.events.member) {
-        return null;
-    }
-    if (allowDefault === undefined) { allowDefault = true; }
-    var rawUrl = member.events.member.getContent().avatar_url;
-    if (rawUrl) {
-        return this._http.getHttpUriForMxc(rawUrl, width, height, resizeMethod);
-    } else if (allowDefault) {
-        return this._http.getIdenticonUri(member.userId, width, height);
-    }
-    return null;
-};
-
-/**
- * Get the avatar URL for a room. <strong>This method is experimental and
- * may change.</strong>
- * @param {module:room.Room} room
- * @param {Number} width The desired width of the thumbnail.
- * @param {Number} height The desired height of the thumbnail.
- * @param {string} resizeMethod The thumbnail resize method to use, either
- * "crop" or "scale".
- * @param {Boolean} allowDefault (optional) Passing false causes this method to
- * return null if the user has no avatar image. Otherwise, a default image URL
- * will be returned.
- * @return {?string} the avatar URL or null.
- */
-MatrixClient.prototype.getAvatarUrlForRoom =
-        function(room, width, height, resizeMethod, allowDefault) {
-
-    if (!room || !room.currentState || !room.currentState.members) {
-        return null;
-    }
-
-    var userId = this.credentials.userId;
-    var members = utils.filter(room.currentState.getMembers(), function(m) {
-        return (m.membership === "join" && m.userId !== userId);
-    });
-
-    if (members[0]) {
-        return this.getAvatarUrlForMember(
-            members[0], width, height, resizeMethod, allowDefault
-        );
-    }
-    return null;
-};
-
-/**
  * Turn an MXC URL into an HTTP one. <strong>This method is experimental and
  * may change.</strong>
  * @param {string} mxcUrl The MXC URL
@@ -1476,7 +1416,9 @@ MatrixClient.prototype.getAvatarUrlForRoom =
  */
 MatrixClient.prototype.mxcUrlToHttp =
         function(mxcUrl, width, height, resizeMethod) {
-    return this._http.getHttpUriForMxc(mxcUrl, width, height, resizeMethod);
+    return contentRepo.getHttpUriForMxc(
+        this.baseUrl, mxcUrl, width, height, resizeMethod
+    );
 };
 
 /**
