+++ lib/models/room-member.js
@@ -3,6 +3,7 @@
  * @module models/room-member
  */
 var EventEmitter = require("events").EventEmitter;
+var contentRepo = require("../content-repo");
 
 var utils = require("../utils");
 
@@ -145,6 +146,38 @@ RoomMember.prototype._updateModifiedTime = function() {
  */
 RoomMember.prototype.getLastModifiedTime = function() {
     return this._modified;
+};
+
+/**
+ * Get the avatar URL for a room member.
+ * @param {string} baseUrl The base homeserver URL
+ * @param {Number} width The desired width of the thumbnail.
+ * @param {Number} height The desired height of the thumbnail.
+ * @param {string} resizeMethod The thumbnail resize method to use, either
+ * "crop" or "scale".
+ * @param {Boolean} allowDefault (optional) Passing false causes this method to
+ * return null if the user has no avatar image. Otherwise, a default image URL
+ * will be returned. Default: true.
+ * @return {?string} the avatar URL or null.
+ */
+RoomMember.prototype.getAvatarUrl =
+        function(baseUrl, width, height, resizeMethod, allowDefault) {
+    if (!this.events.member) {
+        return null;
+    }
+    if (allowDefault === undefined) { allowDefault = true; }
+    var rawUrl = this.events.member.getContent().avatar_url;
+    if (rawUrl) {
+        return contentRepo.getHttpUriForMxc(
+            baseUrl, rawUrl, width, height, resizeMethod
+        );
+    }
+    else if (allowDefault) {
+        return contentRepo.getIdenticonUri(
+            baseUrl, member.userId, width, height
+        );
+    }
+    return null;
 };
 
 function calculateDisplayName(member, event, roomState) {
