+++ lib/models/room.js
@@ -7,6 +7,7 @@ var EventEmitter = require("events").EventEmitter;
 var RoomState = require("./room-state");
 var RoomSummary = require("./room-summary");
 var utils = require("../utils");
+var contentRepo = require("../content-repo");
 
 /**
  * Construct a new Room.
@@ -38,6 +39,38 @@ function Room(roomId, storageToken) {
     this._redactions = [];
 }
 utils.inherits(Room, EventEmitter);
+
+/**
+ * Get the avatar URL for a room if one was set.
+ * @param {String} baseUrl The homeserver base URL. See {@link MatrixClient#getHomeserverUrl()}.
+ * @param {Number} width The desired width of the thumbnail.
+ * @param {Number} height The desired height of the thumbnail.
+ * @param {string} resizeMethod The thumbnail resize method to use, either
+ * "crop" or "scale".
+ * @param {boolean} allowDefault True to allow an identicon for this room if an
+ * avatar URL wasn't explicitly set. Default: true.
+ * @return {?string} the avatar URL or null.
+ */
+Room.prototype.getAvatarUrl = function(baseUrl, width, height, resizeMethod, allowDefault) {
+    var roomAvatarEvent = this.currentState.getStateEvents("m.room.avatar", "");
+    if (!roomAvatarEvent) {
+        return null;
+    }
+    if (allowDefault === undefined) { allowDefault = true; }
+    var mainUrl = roomAvatarEvent.getContent().url;
+    if (mainUrl) {
+        return contentRepo.getHttpUriForMxc(
+            baseUrl, mainUrl, width, height, resizeMethod
+        );
+    }
+    else if (allowDefault) {
+        return contentRepo.getIdenticonUri(
+            baseUrl, member.userId, width, height
+        );
+    }
+
+    return null;
+};
 
 /**
  * Get a member from the current room state.
